tail: large refactoring and cleanup of the tail code base. See also #3905 for details
